Remove comments, remove the limit on # bombs. Start grader mode

+++ plane.cpp
@@ -28,11 +28,11 @@ class Plane : public Displayable {
 			}
 
 			XFillArc(xInfo.display, xInfo.window, xInfo.gc[2],
-				(double)x*xInfo.width/800, (double)y*xInfo.height/600, 
-				(double)width*xInfo.width/800, (double)height*xInfo.height/600, 0, 360*64);
+				x * xInfo.width/800, y * xInfo.height/600, 
+				width * xInfo.width/800, height * xInfo.height/600, 0, 360*64);
 			XFillRectangle(xInfo.display, xInfo.window, xInfo.gc[2], 
-				(double)(x-15)*xInfo.width/800, (double)(y+10)*xInfo.height/600, 
-				(double)30*xInfo.width/800, (double)10*xInfo.height/600);
+				(x-15) * xInfo.width/800, (y+10) * xInfo.height/600, 
+				30 * xInfo.width/800, 10 * xInfo.height/600);
 		}
 
 		int getX() {
